+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.IntegrationTests/Transport/TestImplementations/TransportTestInboundMapper.cs
@@ -16,7 +16,7 @@ using Energinet.DataHub.MarketRoles.Application.Transport;
 using Energinet.DataHub.MarketRoles.Contracts;
 using Energinet.DataHub.MarketRoles.Infrastructure.Transport.Protobuf;
 
-namespace Energinet.DataHub.MarketRoles.IntegrationTests.Send
+namespace Energinet.DataHub.MarketRoles.IntegrationTests.Transport.TestImplementations
 {
     public class TransportTestInboundMapper : ProtobufInboundMapper<TestMessage>
     {
