+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.IntegrationTests/Transport/TestImplementations/TransportTestRecord.cs
@@ -12,23 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 using Energinet.DataHub.MarketRoles.Application.Transport;
 
-namespace Energinet.DataHub.MarketRoles.IntegrationTests.Send
+namespace Energinet.DataHub.MarketRoles.IntegrationTests.Transport.TestImplementations
 {
-    public class InProcessChannel : Channel
-    {
-        private byte[] _writtenBytes;
-
-        public byte[] GetWrittenBytes() => _writtenBytes ?? throw new InvalidOperationException("Write bytes before getting them.");
-
-        public override async Task WriteAsync(byte[] data, CancellationToken cancellationToken = default)
-        {
-            _writtenBytes = data;
-            await Task.CompletedTask.ConfigureAwait(false);
-        }
-    }
+    public record TransportTestRecord(string TestProperty) :
+    IOutboundMessage,
+    IInboundMessage;
 }
